Added a separate layout directory
