+++ src/layout/variables/MainPageConstants.ts
